+++ app/src/androidTest/java/app/envelop/AppTest.kt
@@ -1,7 +1,7 @@
 package app.envelop
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import app.envelop.test.AppHelper.application
+import app.envelop.test.AppHelper.app
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -11,7 +11,7 @@ class AppTest {
 
   @Test
   fun mode() {
-    assertEquals(application.mode, App.Mode.Test)
+    assertEquals(app.mode, App.Mode.Test)
   }
 
 }