+++ app/src/androidTest/java/app/envelop/test/AppHelper.kt
@@ -5,5 +5,6 @@ import app.envelop.App
 
 object AppHelper {
   val context get() = InstrumentationRegistry.getInstrumentation().targetContext!!
-  val application get() = context.applicationContext as App
+  val app get() = context.applicationContext as App
+  val appComponent get() = app.component
 }